+++ CHANGES.md
@@ -17,6 +17,7 @@ Features
   ```
   (CONNECTED) /> ls / watch=true
   ```
+- new command `time`: allows time the execution of the given commands
 
 1.1.0 (2015-06-17)
 --------------------